% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BIC_lowmem}
\alias{BIC_lowmem}
\title{Calculates BIC for a given clustering.}
\usage{
BIC_lowmem(data, centres)
}
\arguments{
\item{data}{matrix of data}

\item{centres}{matrix of cluster means}
}
\value{
BIC value
}
\description{
Calculates BIC for a given clustering.
}

