% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tkmeans_lowmem}
\alias{tkmeans_lowmem}
\title{Performs trimmed k-means clustering algorithm on a matrix of data. Each row is an observation.}
\usage{
tkmeans_lowmem(M, n_cluster, alpha, nstart, iter, tol)
}
\arguments{
\item{M}{matrix of data to be clustered}

\item{n_cluster}{number of clusters}

\item{alpha}{proportion of data to be trimmed}

\item{nstart}{number of restarts}

\item{iter}{maximum number of iterations}

\item{tol}{convergence criteria}
}
\value{
matrix of cluster centres
}
\description{
Performs trimmed k-means clustering algorithm on a matrix of data. Each row is an observation.
}

