% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tmeansClust_lowmem}
\alias{tmeansClust_lowmem}
\title{Allocates each Row (observation) in data to the nearest cluster centre.}
\usage{
tmeansClust_lowmem(data, centres)
}
\arguments{
\item{data}{matrix of data to be clustered}

\item{centres}{set of cluster means}
}
\value{
vector of cluster allocations
}
\description{
Allocates each Row (observation) in data to the nearest cluster centre.
}

