% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scale_lowmem}
\alias{scale_lowmem}
\title{Recales each row to have mean 0 and sd 1.}
\usage{
scale_lowmem(M)
}
\arguments{
\item{M}{matrix of data}
}
\value{
first row mean shifts, second row sd so that scaling can be undone. Scales original matrix in place
}
\description{
Recales each row to have mean 0 and sd 1.
}

